requirements.txt
spotipy~=2.18.0
requests~=2.25.1
pandas~=1.2.5
matplotlib~=3.4.2
seaborn~=0.11.1
Pillow~=8.2.0